Not hide console on Windows during development
On Windows the console window is only hidden if it has been active for
less than a second.
This heuristic assumes that if you 'go run' from the console, the
console window will always have been open for >1 sec so it is not
hidden.
When the user double clicks the executable in Windows, the console will
show up just before the window itself, this should always be <1 sec,
and the console is hidden in that case.