Remove BlurText, remove SDL2, use Go font
The font rendering was completely overhauled. It now uses Go-Mono.ttf,
one of the Go fonts. It uses a bitmap font, generated by a C program
which uses stb_truetype, for desktop (Windows and GLFW, SDL2 was
dropped). For WASM, it uses the TTF font file directly. Everything was
tweaked to make the desktop and WASM texts look as similar as possible.

With the new text rendering, there is no crisp, pixelated text anymore,
so the BlurText function was removed from Window.

The SDL2 backend was removed altogether since GLFW has been the default
for long, was tested on Mac, and our SDL2 backend was not up-to-date
anymore.